pyproject.toml
[project]
name = "digibrain"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot[ext]~=22.6",
    "python-dotenv~=1.1",
    "typer~=0.15",
    "fastapi~=0.115",
    "uvicorn~=0.34",
    "jinja2~=3.1",
    "python-multipart~=0.0.20",
    "prefect>=3.0",
    "anthropic>=0.40",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tools", "scripts", "bot"]

[project.scripts]
applyops = "tools.applyops.cli:app"
gmail = "tools.gmail:main"
agent-bot = "bot.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
